libcurl.hpp
